+++ src/main/java/io/bux/matchingengine/engine/OrderTree.java
@@ -1,15 +1,16 @@
 package io.bux.matchingengine.engine;
 
+import java.math.BigDecimal;
 import java.util.*;
 
 public class OrderTree {
-    private final Map<String, TreeMap<Double, LinkedList<Order>>> assets = new TreeMap<>();
+    private final Map<String, TreeMap<BigDecimal, LinkedList<Order>>> assets = new TreeMap<>();
     private final HashMap<Long, Order> orderMap = new HashMap<>();
 
     public void addOrder(Order order) {
         String asset = order.getAsset();
         Long orderId = order.getOrderId();
-        Double price = order.getPrice();
+        BigDecimal price = order.getPrice();
 
 
         orderMap.put(orderId, order);
@@ -19,7 +20,7 @@ public class OrderTree {
             } else {
                 LinkedList<Order> priceList = new LinkedList<>();
                 priceList.add(order);
-                TreeMap<Double, LinkedList<Order>> orderTree = assets.get(asset);
+                TreeMap<BigDecimal, LinkedList<Order>> orderTree = assets.get(asset);
                 orderTree.put(order.getPrice(), priceList);
             }
         } else {
@@ -30,7 +31,7 @@ public class OrderTree {
     private void saveOrder(Order order) {
         LinkedList<Order> priceList = new LinkedList<>();
         priceList.add(order);
-        TreeMap<Double, LinkedList<Order>> orderTree = new TreeMap<>();
+        TreeMap<BigDecimal, LinkedList<Order>> orderTree = new TreeMap<>();
         orderTree.put(order.getPrice(), priceList);
         assets.put(order.getAsset(), orderTree);
     }
@@ -38,7 +39,7 @@ public class OrderTree {
     public void deleteOrder(Long orderId) {
         if (orderMap.containsKey(orderId)) {
             Order order = orderMap.get(orderId);
-            double price = order.getPrice();
+            BigDecimal price = order.getPrice();
             String asset = order.getAsset();
             if (assets.containsKey(asset)) {
                 if (assets.get(asset).containsKey(price)) {
@@ -61,12 +62,12 @@ public class OrderTree {
         return List.copyOf(assets.get(asset).firstEntry().getValue());
     }
 
-    public double getLowestPrice(String asset) {
-        return assets.isEmpty() || assets.get(asset).isEmpty() ? Double.MAX_VALUE : assets.get(asset).firstKey();
+    public BigDecimal getLowestPrice(String asset) {
+        return assets.isEmpty() || assets.get(asset).isEmpty() ? BigDecimal.valueOf(Double.MAX_VALUE) : assets.get(asset).firstKey();
     }
 
-    public double getHighestPrice(String asset) {
-        return assets.isEmpty() || assets.get(asset).isEmpty() ? Double.MAX_VALUE : assets.get(asset).lastKey();
+    public BigDecimal getHighestPrice(String asset) {
+        return assets.isEmpty() || assets.get(asset).isEmpty() ? BigDecimal.valueOf(Double.MAX_VALUE) : assets.get(asset).lastKey();
     }
 
     public List<Order> getMaxPriceList(String asset) {
